tests/wpt/web-platform-tests/app-history/app-history-entry/entries-in-new-javascript-url-iframe.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="javascript:'foo'"></iframe>

<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);

  assert_not_equals(i.contentWindow.appHistory.current, null);
  assert_array_equals(i.contentWindow.appHistory.entries(), [i.contentWindow.appHistory.current]);

  assert_equals(i.contentWindow.appHistory.current.url, "about:blank");
}, "AppHistory entries() and current should be set in a new javascript: URL iframe");
</script>
